+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ To be released
 * Added ``client.relations.block`` and ``client.relations.unblock`` for blocking/unblocking users.
 * Implemented ``/api/games/export/imports`` under
   ``client.games.export_imported``.
+* Added ``client.account.get_timeline`` to get current logged user timeline.
 
 Thanks to all the contributors who helped to this release:
 - @hsheth2
@@ -19,6 +20,7 @@ Thanks to all the contributors who helped to this release:
 - @MrElyazid
 - @gameroman
 - @JAMoreno-Larios
+
 
 v0.14.0 (2025-08-26)
 --------------------